+++ Assets/Examples/DefaultNodes/Nodes/DynamicPortGeneration/NamerNode.cs
@@ -3,10 +3,9 @@ using System.Collections.Generic;
 using UnityEngine;
 using GraphProcessor;
 using System.Linq;
-using static SequenceName;
 
-[System.Serializable, NodeMenuItem("Custom/ConditionalNameNode")]
-public class ConditionalNameNode : DynamicNodeWithOutput<ConditionalName>
+[System.Serializable, NodeMenuItem("Custom/ProxiedInputsNode")]
+public class NamerNode : DynamicNodeWithOutput<Namer>
 {
     public override string name => "ConditionalNameNode";
 }
