--- Assets/Testing/BaseIsConditional.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9947347622c831651bc3a306795104fa
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
