--- Assets/Testing/DelegateEvent.VarObjectEventArgs.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1d0825e47caa8faf0b509d6413c06002
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
