--- Assets/Testing/DelegateEvent.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b336ffaa1cc79bc929414981fe23fab5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
