--- Assets/Testing/ExpandableSOAttribute.cs
@@ -1,15 +1,0 @@
-using System;
-using UnityEditor;
-using UnityEngine;
-
-/// <summary>
-/// Attribute takes the insides of a set ScriptableObject and display it.
-/// </summary>
-public class ExpandableSOAttribute : PropertyAttribute
-{
-
-    /// <summary>
-    /// Required implementation of the <see cref="T:ExpandSOAttribute"/> interface.
-    /// </summary>
-    public ExpandableSOAttribute() { }
-}
