--- Assets/Testing/ExpandableSOAttribute.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bf7bd1d3b461c28ca869d43343391f92
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
