--- Assets/Testing/MyInputAttribute.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0189acdd3b39929d6a8b0a57c3e507bd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
