--- Assets/Testing/PortUpdaterNode.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 67b3cddba608059109efbb060f6504fe
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
