--- Assets/Testing/SharedVariable.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4ace1829c0b27489fa70482f81ddbed7
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
