--- Assets/Testing/ValueChangedCallbackAttribute.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e9b5389fba495b15191a31e041bfedba
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
