+++ Assets/com.alelievr.NodeGraphProcessor/Runtime/Utils/FieldInfoExtension.cs
@@ -1,9 +1,7 @@
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
-using UnityEngine;
 
 namespace GraphProcessor
 {
