+++ Assets/com.alelievr.NodeGraphProcessor/Runtime/Utils/SerializedEdgeExtension.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: c2ed7d6a9e49992bca5a969a1db59655
+guid: 1b6986467dd851f8b8153d3bf6b93994
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
